backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
langchain==0.1.6
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
tiktoken==0.5.2
pinecone-client==3.0.2
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.2
bleach==6.1.0
